requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework
markdown
django-filter
djangorestframework_simplejwt
pyyaml